fix(isaak): resolver conflicto de ruta /settings — mover a route group (workspace)
PR #91 creo app/settings/ (server wrapper + IsaakSettingsClient con templates/certs)
dejando tambien app/(workspace)/settings/page.tsx original. Next.js rechaza
dos paginas que resuelven al mismo path /settings.

Fix: mueve server wrapper + IsaakSettingsClient al route group (workspace) para
que /settings tenga el layout con sidebar. Elimina app/settings/ raiz.
